requirements.txt
click
mysql-connector-python
